requirements.txt
pyTelegramBotAPI
Pillow